# src/main/webapp/api-docs/openapi.json
{
    "openapi": "3.0.1",
    "servers": [
        { "url": "/msa-api-gateway/gateway", "description": "API Gateway" }
    ],
...
    "paths": {
        "/agency/create/airline": {
            "post": {
                "tags": [
                    "Agencia"
                ],
                "summary": "Crear reserva de aviÃ³n",
                "operationId": "createAirlineAndHotelReservation",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReservationAirlineRequestDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reserva de aviÃ³n creada"
                    }
                }
            }
        },
        "/agency/create/hotel-airline": {
            "post": {
                "tags": [
                    "Agencia"
                ],
                "summary": "Crear reserva combinada hotel + aviÃ³n",
                "operationId": "createAirlineAndHotelReservation_1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateAirlineAndHotelReservationDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reserva combinada creada"
                    }
                }
            }
        },
        "/agency/create/hotel": {
            "post": {
                "tags": [
                    "Agencia"
                ],
                "summary": "Crear reserva de hotel",
                "operationId": "createHotelReservation",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateHotelBookingDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reserva de hotel creada"
                    }
                }
            }
        },
        "/agency/delete/hotel-airline/{reservationId}/{bookingId}": {
            "delete": {
                "tags": [
                    "Agencia"
                ],
                "summary": "Eliminar reserva combinada hotel + aviÃ³n",
                "operationId": "deleteAirlineAndHotelReservation",
                "parameters": [
                    {
                        "name": "reservationId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "bookingId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reserva combinada eliminada"
                    }
                }
            }
        },
        "/agency/delete/airline/{reservationId}": {
            "delete": {
                "tags": [
                    "Agencia"
                ],
                "summary": "Eliminar reserva de aviÃ³n",
                "operationId": "deleteAirlineReservation",
                "parameters": [
                    {
                        "name": "reservationId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reserva de aviÃ³n eliminada"
                    }
                }
            }
        },
        "/agency/delete/hotel/{bookingId}": {
            "delete": {
                "tags": [
                    "Agencia"
                ],
                "summary": "Eliminar reserva de hotel",
                "operationId": "deleteHotelBooking",
                "parameters": [
                    {
                        "name": "bookingId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reserva de hotel eliminada"
                    }
                }
            }
        },
        "/agency/hotel-airline/{hotelId}/{airlineId}": {
            "get": {
                "tags": [
                    "Agencia"
                ],
                "summary": "Obtener viaje por ID de hotel y aerolinea",
                "operationId": "getTravelByHotelAndAirline",
                "parameters": [
                    {
                        "name": "hotelId",
                        "in": "path",
                        "description": "Identificador del hotel",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "airlineId",
                        "in": "path",
                        "description": "Identificador de la aerolinea",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Viaje encontrado"
                    },
                    "404": {
                        "description": "Viaje no encontrado"
                    }
                }
            }
        },
        "/agency/travel/{id}": {
            "get": {
                "tags": [
                    "Agencia"
                ],
                "summary": "Obtener viaje por ID",
                "operationId": "getTravelById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identificador del viaje",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Viaje encontrado"
                    },
                    "404": {
                        "description": "Viaje no encontrado"
                    }
                }
            }
        },
        "/agency/travel/user/{idUser}": {
            "get": {
                "tags": [
                    "Agencia"
                ],
                "summary": "Obtener viajes por ID de usuario",
                "operationId": "getTravelByIdUser",
                "parameters": [
                    {
                        "name": "idUser",
                        "in": "path",
                        "description": "Identificador del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de viajes obtenida"
                    }
                }
            }
        },
        "/agency/list": {
            "get": {
                "tags": [
                    "Agencia"
                ],
                "summary": "Obtener lista de viajes de Hotel con Aerolinea",
                "operationId": "getTravelList",
                "parameters": [
                    {
                        "name": "countryOrigin",
                        "in": "query",
                        "description": "PaÃ­s de origen (ISO-3166 alfa-2 o nombre)",
                        "schema": {
                            "type": "string"
                        },
                        "example": "ES"
                    },
                    {
                        "name": "countryDestination",
                        "in": "query",
                        "description": "PaÃ­s de destino (ISO-3166 alfa-2 o nombre)",
                        "schema": {
                            "type": "string"
                        },
                        "example": "US"
                    },
                    {
                        "name": "cityOrigin",
                        "in": "query",
                        "description": "Ciudad de origen",
                        "schema": {
                            "type": "string"
                        },
                        "example": "Madrid"
                    },
                    {
                        "name": "cityDestination",
                        "in": "query",
                        "description": "Ciudad de destino",
                        "schema": {
                            "type": "string"
                        },
                        "example": "New York"
                    },
                    {
                        "name": "dateOrigin",
                        "in": "query",
                        "description": "Fecha de salida (formato ISO: yyyy-MM-dd)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        },
                        "example": "2025-10-01"
                    },
                    {
                        "name": "hotelName",
                        "in": "query",
                        "description": "Nombre del hotel",
                        "schema": {
                            "type": "string"
                        },
                        "example": "Gran VÃ­a Palace"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de viajes obtenida"
                    }
                }
            }
        },
        "/agency/update/hotel-airline": {
            "put": {
                "tags": [
                    "Agencia"
                ],
                "summary": "Actualizar reserva combinada hotel + aviÃ³n",
                "operationId": "updateAirlineAndHotelReservation",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateReservationBookingDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reserva combinada actualizada"
                    }
                }
            }
        },
        "/agency/update/airline": {
            "put": {
                "tags": [
                    "Agencia"
                ],
                "summary": "Actualizar reserva de aviÃ³n",
                "operationId": "updateAirlineReservation",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateReservationDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reserva de aviÃ³n actualizada"
                    }
                }
            }
        },
        "/agency/update/hotel": {
            "put": {
                "tags": [
                    "Agencia"
                ],
                "summary": "Actualizar reserva de hotel",
                "operationId": "updateHotelBooking",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateHotelBookingDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reserva de hotel actualizada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "FlightInstanceSeatsDTO": {
                "type": "object",
                "properties": {
                    "idFlightInstance": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "numberSeats": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "ReservationAirlineRequestDTO": {
                "type": "object",
                "properties": {
                    "dni": {
                        "type": "string"
                    },
                    "idUser": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "listIdFlightInstanceSeats": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FlightInstanceSeatsDTO"
                        }
                    }
                }
            },
            "CreateAirlineAndHotelReservationDTO": {
                "type": "object",
                "properties": {
                    "idUser": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "customer": {
                        "$ref": "#/components/schemas/CustomerDTO"
                    },
                    "infoReservation": {
                        "$ref": "#/components/schemas/CreateAirlineAndHotelReservationInfoDTO"
                    }
                }
            },
            "CreateAirlineAndHotelReservationInfoDTO": {
                "type": "object",
                "properties": {
                    "idCountry": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "startDate": {
                        "type": "string"
                    },
                    "endDate": {
                        "type": "string"
                    },
                    "numberOfNights": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "peopleNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "withBreakfast": {
                        "type": "boolean"
                    },
                    "flightInstances": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FlightInstanceSeatsDTO"
                        }
                    },
                    "roomIds": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                }
            },
            "CustomerDTO": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "dni": {
                        "type": "string"
                    }
                }
            },
            "CreateHotelBookingDTO": {
                "type": "object",
                "properties": {
                    "customer": {
                        "$ref": "#/components/schemas/CustomerDTO"
                    },
                    "startDate": {
                        "type": "string"
                    },
                    "endDate": {
                        "type": "string"
                    },
                    "numberOfNights": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "withBreakfast": {
                        "type": "boolean"
                    },
                    "peopleNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "userId": {
                        "type": "string"
                    },
                    "roomsIds": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                }
            },
            "UpdateReservationBookingDTO": {
                "type": "object",
                "properties": {
                    "bookingId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "customerDNI": {
                        "type": "string"
                    },
                    "startDate": {
                        "type": "string"
                    },
                    "endDate": {
                        "type": "string"
                    },
                    "numberOfNights": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "withBreakfast": {
                        "type": "boolean"
                    },
                    "peopleNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "roomsInfo": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    "idReservation": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "idTravel": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "flightInstanceSeats": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FlightInstanceSeatsDTO"
                        }
                    }
                }
            },
            "UpdateReservationDTO": {
                "type": "object",
                "properties": {
                    "idReservation": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "idTravel": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "flightInstanceSeats": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FlightInstanceSeatsDTO"
                        }
                    }
                }
            },
            "UpdateHotelBookingDTO": {
                "type": "object",
                "properties": {
                    "bookingId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "travelId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "customerDNI": {
                        "type": "string"
                    },
                    "startDate": {
                        "type": "string"
                    },
                    "endDate": {
                        "type": "string"
                    },
                    "numberOfNights": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "withBreakfast": {
                        "type": "boolean"
                    },
                    "peopleNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "roomsInfo": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                }
            },
...
        }
    }
}